Output of observed mixture props precision
added to libutils a gsl error handler that throws exceptions, a set of gsl exception classes to throw and an InfiniteGradient exception.

modified myexp function to use new error handler.

LambdaGradient in PopHapMix throws InfiniteGradient if it catches an underflow.

added a catch for InfiniteGradient to HamiltonianMonteCarlo.

fixed compilation errors in HapMixIndividual and LogWriter

added some comments to HapMixIndividualCollection