Created a common base-class PedBase for both the Pedigree and Individual classes, containing virtual methods representing all of the public (and some of the protected) methods in Individual and AdmixedIndividual, as a first step towards being able to run on pedigrees the same model that currently runs on individuals.  The base class does not implement any of these methods, but rather throws an exception if any of them (which haven't been overridden) are called.  Since all of the method signatures were taken directly from Individual/AdmixedIndividual, these classes already override all of them; while Pedigree does not, since some of the method signatures do not apply to pedigrees.
Nonetheless, this allows the admixmap model to be run on an array of (pointers to) PedBase's which are in fact either AdmixedIndividual or Pedigree objects, mixed object-by-object.

If there are Pedigree objects in the list, and the options specified to the model cause it to call any of the methods which have not yet been implemented, an exception will be thrown causing execution to abort (with a tidy error message), while if only AdmixedIndividuals are fed to the model, options that cause arbitrary methods to be called may be used.

A new boolean option has been added, use-pedigree-for-individual (accessible in the code as Options::getUsePedForInd()) which will cause single-member Pedigree objects to be constructed for unrelated individuals rather than AdmixedIndividual.  It defaults to false, since there is normally no advantage to this (and in fact it is disadvantageous since the Individual classes implement certain optimizations that Pedigree class does not), but it is extremely useful for testing since it allows the same dataset to be run using either Individual or Pedigree objects (the results should match if the code is working).  Pedigree objects are always constructed for multi-member pedigrees.

At present, not enough of the AdmixedIndividual methods have been implemented in Pedigree for a successful run with use-pedigree-for-individual turned on.  Some of the methods which have been implemented contain code that was copied from Individual or AdmixedIndividual and modified to compile without careful review, sometimes by removing sections of code because it is presumed that they will not be called on pedigrees with certain options turned on (e.g. we are assuming that globalrho is always on).  Most of these methods which are overridden from PedBase and re-implemented from Individual or AdmixedIndividual are gathered into a separate source file, AdmixPedigree.cc (in the admixmap source directory rather than in the base library) in anticipation of creating a separate sub-class of Pedigree, which can then remain a more abstract representation of a pedigree.  The separate class does not as yet exist and the methods are declared in the Pedigree class, in Pedigree.h, to simplify construction of the objects for the moment.

At the same time, some std:vector classes/objects were reparented to bclib::cvector to add conditional range-checking to ease debugging.

svn add src/admixmap/AdmixPedigree.cc src/base/HiddenMarkovModel.new.cc src/base/HiddenMarkovModel.new.h src/base/PedBase.cc src/base/PedBase.h