Add the AdmixtureProportions class and start using it for individuals.
The AdmixtureProportion class uses a cvector< pvector<double> > to store the
admixture proportions. It allows working with each element through operator[],
so that it is possible to use expressions such as Theta[g][k].

This changes the signature of virtual functions Individual::getLogLikelihood() and
UpdateHMMInputs() to use the AdmixtureProportions class rather than flat arrays.
The class can return a flat (read-only) version of its internal vectors, which is
stored within the class and updated when necessary. This allows using the class
without having to change all parts of code that expect to work with flat arrays.
The only places where flat arrays are still passed are HMM::SetStateArrivalProbs()
and AffectedsOnlyTest::Update().

Since the allocation of vectors initialises all elements to zero, using the class
fixes two valgrind errors about conditional jumps depending on uninitialised values
regarding the use of ThetaProposal (at the first iteration) and ThetaBar. These
were occurring because only the values of these arrays for which the corresponding
Theta was strictly positive were set, while the others were left uninitialised.

This also adds versions of gendirichelet() and getDirichletLogDensity_Softmax()
that accept a pvector<double>.