Implement separate admixture proportions for the X chromosome.
This adds the flatXChromosome() method to the AdmixtureProportions class which
returns a flat array adjusted to take into account the odds ratio (psi) for
the X chromosome admixture. When we update the inputs for the Hidden Markov
models (AdmixedIndividual::UpdateHMMInputs()), we can either pass theta or
thetaX depending on the chromosome type.

This includes the implementation of a bisection method to find the root of
a polynomial equation in the computation of the ratios of paternal founders
to maternal founders given the odds ratios and the autosomal admixture
proportions, which is necessary to generate the X chromosome admixture
proportions.

The odds ratios are learned using a random-walk Metropolis-Hastings algorithm
implemented in PopAdmix::UpdateOddsRatios(). In the computation of the
loglikelihoods, we consider the change caused by the odds ratio on the
X chromosome alone (Individual::getLogLikelihoodXChr()) to reduce the
computational effort. At the moment, the Individual's loglikelihood is not
updated as a consequence of accepting a psi proposal.