Use the posterior mean for psi in the posterior means computations.
Before writing the posterior means for the individuals, we have to set the
posterior mean for psi, otherwise we end up performing those computations with
the psi set at the last iteration, which is not what we want. The call to
PopAdmix::StoreOddsRatiosPosteriorMean() ensures that the psi
used by the individuals in those computations is indeed the posterior mean.

This fixes the X chromosome admixture posterior means and the final deviances.