Fix the path that svnversion looks at to get the revision number.
The call to svnversion returns the number of the svn revision for the
currently checked-out working directory (or "exported" if it doesn't
correspond to an svn checkout).

When the builddir is different from srcdir, we should call svnversion over
srcdir (as that's the place where the sources are) rather than over builddir
(which would not correspond to an svn checkout), otherwise we would always
get "exported" rather than the desired revision number.