Make getAdmixtureProps() X chromosome aware and fix AllelicAssocTest.
This changes the signature of getAdmixtureProps() so that we can indicate
whether we want to retrieve the X chromosome admixtures rather than the
autosomal admixtures. The bool parameter has a default value of "false" as
most of the callers aren't aware either about what admixture they work with,
so we leave them unchanged to work with autosomal admixtures.

The only test which loops over chromosomes is the AllelicAssocTest, which is
now adjusted to get the X chromosome admixtures when needed.