Fix floating point exception happening with globalrho=0.
This was caused by (NumOfPsiUpdates % w) in PopAdmix::UpdateOddsRatios(),
where both quantities where zero. But w (the number of steps after which the
stepsize tuner should be updated) should not be zero if the tuner is used,
which is the case when the X chromosome is used even if globalrho=0.