Add HiddenStateSpace::idxMapping() to allow the remapping of the alpha vector.
TransRecursion relies on the fact that the alpha vector has a well-determined
order, so we provide the methods that allow doing a remapping of the indices.

In the remapping we exploit the fact that the ordering of inheritance vectors
is completely irrelevant, as they all have a probability of 0.5. However,
ancestry vectors have to be processed in reverse order of their internal
representation (see to_decimal()).