Implement the TransRecursion algorithm.
This replaces the implementation of HiddenMarkovModel::recursionProbs() with
one based on the TransRecursion algorithm. Since this relies on "slicing" the
probability vector into parts by some simple pointer arithmetics, alpha and
beta must have dimension corresponding to the complete hidden state space
size.

The current implementation is based on a remapping of the vectors
passed to it, so there is a lot of vector copying and reordering
in recursionProbs(), which can be avoided by changing HVIterator.