Allow HiddenStateSpace::Iterator to visit all hidden state space elements.
By default, HiddenStateSpace::Iterator skips the elements with zero emission
probability. However, we may need to traverse all elements independently of
their emission probability.

Therefore we modify the constructor to accept a boolean parameter that
controls whether the zero elements should be skipped or not. The parameter
defaults to true (i.e. those elements are skipped), so that we do not have
to modify the existing users.