Fix the computation of the backwards probability.
The error wass due to the fact that TransRecursion() is specified for the
forward sweep only, while for the backward one we should use the transpose
of the transition probability matrix. It turns out that we can actually use
exactly the same algorithm for the backward pass, if we scale appropriately
the beta vector, as it is done in the individual code.

Therefore, we now scale beta by the stationary distribution before entering
recursionProbs(), and scale it back afterwards. This relies on caching
the stationary distribution (actually, its inverse) when we compute it in
HiddenMarkovModel::computeForwards(). For this to work correctly, we must
force the iterator in that loop to visit also the elements for which the
emission probability is zero.