Avoid the remapping of the probability vectors and keep them sparse.
Instead of reordering the probability vectors inside recursionProbs(), we
change the order in which we consider the ancestries within the ancestry
vectors. In practice, this is accomplished by iterating in reverse order
through the hidden variables for the ancestries.

Moreover, we do not need to use probability vector of full hidden-state space
size throughout the program: this is necessary only in transRecursion().
Therefore, in recursionProbs() we expand the "sparse" vectors into dense
vectors in which the zero elements appear explicitly, and the latter are
the ones passed to transRecursion().

This reverts r2851, since the methods added there are not being used anymore.
It also reverts r2858, since we again use "sparse" vectors for alpha, beta
and condStateProbs, which must be indexed with getNon0Index().