Don't recompute HMM::getCondStateProbsAtLocus() if it's already up to date.
After the conditional state probabilities have been computed, we have to set
dirtyCondStateProbs to false, so we do not recompute them if the forward and
backward probabilities haven't changed.

This fixes a huge performance problem, as getCondStateProbsAtLocus() (after
the TransRecursion changes) had become the most expensive function in the
program.