Enable the calls to SampleHapPairs() also for the affecteds-only test.
Up to r2743 AdmixMapModel::UpdateParameters() would call SampleHapPairs()
unconditionally. In r2744, this was changed so that it would be called only
if any of the association tests were requested. However, since this function
is responsible for updating the allele counts, it must be called also for
the affecteds-only test.

This fixes the affecteds-only test for individuals when fixedallelefreqs=0.