Fix the number of observations given to the Dirichlet sampler.
The number of observations for the Dirichlet sampler was computed as either
the number of individuals (for assortative mating) or its double (for random
mating). This is wrong for pedigrees, and makes the pedigree results to be
dependent on the randommatingmodel option, which shouldn't be the case.

The correct way of counting the number of observations is the following.
Each unrelated individual accounts for either one (for assortative mating)
or two (for random mating) observations; each pedigree always accounts for
its number of founders.