The no-conjugate-update option should not affect the unrelated individuals.
The calls to SampleHiddenStates() and SampleJumpIndicators() from
AdmixIndividualCollection::HMMUpdates() should be executed for unrelated
individuals if at the current iteration we are doing a conjugate update
and globalrho=0.

We were trying to disable those calls for pedigrees by checking the value of
the no-conjugate-update option. However, this had the side effect that the
value of this option would affect the results obtained on datasets with only
unrelated individuals.

We are now able to check whether each element of the dataset is an unrelated
individual or if it is part of a pedigree, so that we can enable those calls
only when needed.