AncestryVector dynamic maximum-founder-gametes.
The hard-coded limit on maximum number of founder-gametes in any pedigree
(AV_MAX_FOUNDER_GAMETES in AncestryVector.h, currently set to 16) had been used
not only to determine the size of a vector (for which it is still used because
we would like to know it at compile-time), but also to allocate the size of the
internal index lookup table, used to translate an index in the "space" of
vectors to its value.  This caused far too much space to be allocated for the
table in cases when K is high (such as 4), and could even cause integer overflow
on the indexes; now the limit is dynamically determined to be the actual largest
occurring value in the input dataset.