Avoid calling addToEProbAt() if the emission probability is zero.
In Pedigree::accumStateInArray() we check whether the emission probability
is zero: in such case, we print out a warning. However, even in such case
we still call HiddenStateSpace::addToEProbAt(), thus adding a zero to the
existing emission probability. This has the undesired side effect of
increasing the count of nonzero elements of the hidden state space.

An alternative fix would be to check explicitly for the value of amt in
addToEProbAt():
   - if ( rv == 0.0 )
   + if ( rv == 0.0 && amt != 0.0 )