Fix bug in computing the locus ancestry probabilities with the X chromosome.
The computation of the locus ancestry probabilities happened in
AdmixedIndividual::getPosteriorMeans(), where the overall sums are divided
by the number of samples. However, getPosteriorMeans() is called by both
WritePosteriorMeans() and WritePosteriorMeansXChr().

This meant that on datasets containing the X chromosome, the locus ancestry
probabilities were divided twice by the number of samples, therefore they
would not sum to 1. To avoid this, we move the code that computes the locus
ancestry probabilities directly inside WritePosteriorMeans() to guarantee
that it will be run just once.