The no-conjugate-update option should not affect the individual code (part 2).
Since pedigrees will always use the random-walk update (which corresponds to
setting the option no-conjugate-update=1), we can pass isPedigree() instead
of getNoConjugateUpdate() in the call to SampleTheta(). This ensures that
when the individual code is run this parameter is always false, rather than
depend on the value of the no-conjugate-update option.

This completes the patch committed in r2875.