Remove the default value from Organism::isHaploid() for the X chromosome.
The presence of the X chromosome may affect the number of gametes modelled
in a pedigree member. For a male founder with genotyped data, we would model
two gametes on the autosomes and one gamete on the X chromosome. For this
reason, we need to know the locus we are referring to when testing for
haploidness.

In five places (marked with XXX) we have no knowledge of the locus under
consideration, so I pass the false value so that the code compiles.

Note that this is Marco's work, but I (David) am checking it into SVN
from my git repo.