Start using the AdmixtureProportions class in the pedigree code.
This introduces the new AdmixType typedef and uses it in a number of places
where we are passing around the object containing the admixture proportions
(previously a ThetaType object). This affects the type of objects stored
in the ProposalRingBuffer and the TransProbCache.

This is the first step in getting the correct admixture proportions for the
X chromosome on pedigrees.