Move some common code inside HiddenMarkovModel::recursionProbs().
This helps consolidating some code that is present in both computeForwards()
and computeBackwards() by moving it into recursionProbs().

The scaling of theta by 1-f has been moved as it will allow for cleaner code
for the X chromosome, and for the foreseable future there are no plans to
implement the caching of the scaled-theta as suggested by the comment.