Allow for allele frequency of zero in the individual code.
This fixes discrepancy in emission probabilities between individual and
pedigree codes for the case in which we use the true allele frequencies,
that is the allele probabilities are 0 and 1.

Also, we now throw in case we read a negative values, while before this was
silently set to a positive value.