Allow the printing out of the random numbers generated.
This adds a debugging printout to myrand(), gennor() and gengam(), which are
the most used random number functions in admixmap.

This is most useful in association to other debugging printouts (such as
DEBUG_UPDATE_ODDS_RATIOS, DEBUG_UPDATE_GLOBAL_SUM_INTENSITIES and
DEBUG_TH_PROP) so that it becomes easier to localise where the calls to these
functions happen.