The code doesn't actually support genotypes containing Mendelian errors.
We now refuse to run if exclude-mendelian-errors=0 is set, as it causes an
assertion in HiddenMarkovModel::computeForwards() in that the normalization
factor computed there is zero. See also the comment towards the end of
InputAdmixData::finishConstructing().

This also removes an extra newline in the messages reported when a Mendelian
error is encountered.