Improve the output displayed to the screen.
We set a fixed format for the output of floating point numbers so that we
have less ragged columns. This also corrects the computation of the width
for the iteration number.

The output of the psi parameter is reduced to 3 decimal places to make it
more distinguishable from the other numbers and to avoid printing out
1.000000 for the first odds ratio.