Make Options::checkOptions() return a bool and actually check it.
The return value of Options::checkOptions() was not actually used in setting
the value of badOptions in AdmixOptions::checkOptions(), which could lead to
running the program when it should have stopped.

This also uniforms some of the messages when some of the options are ignored
(it reports them as warnings, since the program actually continues to run).