Speed up AdmixedIndividual::UpdateScoreTests() on the X chromosome.
The admixture proportions do not depend on the locus of a given chromosome,
so there is no point in extracting the flat array of admixtures at each
locus. This observation is particularly important on the X chromosome, since
each call to flatXChromosome() implies the recomputation of the X-chromosome
admixture proportions from scratch.

On a simulated dataset with 100 unrelated individuals and 861 loci (501 of
which on the X chromosome), the running time goes from 32s to 26s, with the
number of calls to flatXChromosome() from UpdateScoreTests() going from
5,010,000 to 10,000.