Speed up Pedigree::accumAOScore() on the X chromosome.
This is the pedigree equivalent of r3013. Note that with this change we are
going to have an unnecessary call getTheta(PsiType&) if the data doesn't
contain the X chromosome, since we don't have a way of knowing that in the
pedigree code. However, the reduction in number of calls to getTheta() more
than compensates for that.

The same dataset (100 unrelated individuals, 861 total loci, 501 of which
on the X chromosome, 200 iterations, 100 burnin), goes from 100s to 94s,
with the number of calls to getTheta() and getTheta(PsiType&) reduced from
1,781,250 and 2,493,750 to 58,050 and 93,750, respectively.