Use the values for f and g computed and cached in tpCache.
The g factors are computed in the TransProbCache constructor, and they never
change throughout the iterations, as they only depend on the locus distance.
The f factors depend on rho: the call to TransProbCache::reComputeFactors()
from setRho() ensures that the f factors are always up-to-date with respect
to the most recent value of rho.

This introduces TransProbCache::getF() and getG() to retrieve the cached
values, so computeF() and computeG() do not need to remain in the public
interface.

With this change, the data set with 100 unrelated individuals, 861 loci run
with the pedigree code for 200 iterations goes from 94s to 89s.