Fix the compilation from a source package.
In an admixmap source package created with 'make admixmap-srcdist' the base,
bclib and admixmap directories are at the top level, as opposed to under the
src directory like in the svn repository.

The Makefile.am file in admixmap/ hardcoded 'src' in the paths for the
libraries and so it could not find libbase.a.