Parallelize the for loops in PopAdmix::UpdateOddsRatios().
There was a significant loss of parallel performance on the amass data with
the X chromosome. UpdateOddsRatios() was running in serial when looping over
the pedigrees in order to accept or reject the psi proposal.

This adds the necessary parallel regions around those loops, and also around
the loops in which we compute the log-likelihoods (they didn't seem to cause
any slowdown before, and they were being run in parallel even without being
explicitly marked as such).

With this change, running the amass data for on 8 cores for two iterations
takes 6m14s instead of 8m54s.