No need to check for a C compiler.
We don't need to check if we have a C compiler for gsl: that check belongs
to the gsl library (if we decide to compile it ourselves), or we may install
a pre-packaged gsl library, in which case we don't need a C compiler at all.

This should speed up configure a bit.