Reduce the allocations made by CopyNumberAssocTest::Update().
This function used to allocate both a vector X and its copy Xcopy. Xcopy was
necessary because X is scaled when computing the expectaton of the score.
However, Xcopy is unnecessary if we move the scaling of X when X itself is
not needed anymore (that is, after computing the uncorrected info).