The order of the source files in the makefile affects the speed.
In r3048 the makefiles were rewritten to be in alphabetical order. However,
that causes a slowdown in the amass data set by about 1 minute over just 2
iterations (see ticket #137). This is the smallest patch that restores the
previous performance: it turns out the AdmixFreqs should be listed after
AdmixPedigree, otherwise we see a slowdown.